fix: Move error handling emails to Middleware
Now that the POC works, we can implement error notifications properly
